news_scraper/requirements.txt
scrapy
tap_news_utils
python-dateutil
newspaper3k
